add bianji

+++ src/main/java/life/lxd/community/dto/QuestionDTO.java
@@ -8,13 +8,13 @@ import lombok.Data;
  */
 @Data
 public class QuestionDTO {
-    private Long id;
+    private Integer id;
     private String title;
     private String description;
     private String tag;
     private Long gmtCreate;
     private Long gmtModified;
-    private Long creator;
+    private Integer creator;
     private Integer viewCount;
     private Integer commentCount;
     private Integer likeCount;
